✨ Feat:db항목 조회 메소드 추가
월별, 일별 항목 조회 메소드 추가

+++ app/src/main/java/com/example/capstone_donworry/DBHelper.java
@@ -30,6 +30,7 @@ public class DBHelper extends SQLiteOpenHelper {
 
     public DBHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
+        this.context = context;
     }
 
     @Override
@@ -52,17 +53,20 @@ public class DBHelper extends SQLiteOpenHelper {
         onCreate(db);
     }
 
-    // db에 저장되어 있는 모든 항목 정보
+    // db에 저장되어 있는 같은 달 항목 정보
     @SuppressLint("Range")
-    public List<AmountItem> getAllItems() {
+    public List<AmountItem> getMonthItems(String date) {
         List<AmountItem> items = new ArrayList<>();
-        String selectQuery = "SELECT * FROM " + item_TABLE_NAME;
+        String selectQuery = "SELECT * FROM " + item_TABLE_NAME +
+                " WHERE strftime('%Y', " + COL_DATE + ") = strftime('%Y', ?) " +
+                " AND strftime('%m', " + COL_DATE + ") = strftime('%m', ?)";
         SQLiteDatabase db = this.getReadableDatabase();
-        Cursor cursor = db.rawQuery(selectQuery, null);
+        Cursor cursor = db.rawQuery(selectQuery, new String[]{date, date});
 
         if (cursor.moveToFirst()) {
             do {
                 AmountItem amountItem = new AmountItem();
+                amountItem.setUid(cursor.getInt(cursor.getColumnIndex(UID)));
                 amountItem.setContent(cursor.getString(cursor.getColumnIndex(COL_CONTENT)));
                 amountItem.setDate(cursor.getString(cursor.getColumnIndex(COL_DATE)));
                 amountItem.setCard(cursor.getString(cursor.getColumnIndex(COL_CARD)));
@@ -76,18 +80,48 @@ public class DBHelper extends SQLiteOpenHelper {
         db.close();
         return items;
     }
+    // TODO: 월 별 합계 구하기
+
+
+    // db에 저장되어 있는 같은 날짜 항목 정보
+    @SuppressLint("Range")
+    public List<AmountItem> getDateItems(String date) {
+        List<AmountItem> items = new ArrayList<>();
+        String selectQuery = "SELECT * FROM " + item_TABLE_NAME + " WHERE " + COL_DATE + " = ?";
+        SQLiteDatabase db = this.getReadableDatabase();
+        Cursor cursor = db.rawQuery(selectQuery, new String[]{date});
+
+        if (cursor.moveToFirst()) {
+            do {
+                AmountItem amountItem = new AmountItem();
+                amountItem.setUid(cursor.getInt(cursor.getColumnIndex(UID)));
+                amountItem.setContent(cursor.getString(cursor.getColumnIndex(COL_CONTENT)));
+                amountItem.setDate(cursor.getString(cursor.getColumnIndex(COL_DATE)));
+                amountItem.setCard(cursor.getString(cursor.getColumnIndex(COL_CARD)));
+                amountItem.setBank(cursor.getString(cursor.getColumnIndex(COL_BANK)));
+                amountItem.setCategory(cursor.getString(cursor.getColumnIndex(COL_CATEGORY)));
+                amountItem.setAmount(cursor.getString(cursor.getColumnIndex(COL_AMOUNT)));
+                items.add(amountItem);
+            } while (cursor.moveToNext());
+        }
+        cursor.close();
+        db.close();
+        return items;
+    }
+    // TODO: 일별 합계 구하기
 
     // db에 저장되어 있는 특정 항목 정보
     @SuppressLint("Range")
     public AmountItem getItem(AmountItem amountItem) {
         AmountItem item = new AmountItem();
-        String selectQuery = "SELECT * FROM " + item_TABLE_NAME + " WHERE " + COL_CONTENT + " = ? AND " + COL_AMOUNT + " = ? AND " + COL_DATE + " = ?";
-        String[] selectionArgs = {amountItem.getContent(), amountItem.getAmount(), amountItem.getDate()};
+        String selectQuery = "SELECT * FROM " + item_TABLE_NAME + " WHERE " + COL_CONTENT + " = ? AND " + COL_DATE + " = ? AND " + UID + " = ?";
+        String[] selectionArgs = {amountItem.getContent(), amountItem.getDate(), String.valueOf(amountItem.getUid())};
 
         SQLiteDatabase db = this.getReadableDatabase();
         Cursor cursor = db.rawQuery(selectQuery, selectionArgs);
 
         if (cursor.moveToFirst()) {
+            amountItem.setUid(cursor.getInt(cursor.getColumnIndex(UID)));
             item.setContent(cursor.getString(cursor.getColumnIndex(COL_CONTENT)));
             item.setDate(cursor.getString(cursor.getColumnIndex(COL_DATE)));
             item.setCard(cursor.getString(cursor.getColumnIndex(COL_CARD)));
@@ -101,19 +135,33 @@ public class DBHelper extends SQLiteOpenHelper {
     }
 
     // 새로운 item 추가
-    public void addItem(AmountItem amountItem) {
-        SQLiteDatabase db = this.getWritableDatabase();
-        ContentValues values = new ContentValues();
-        values.put(COL_CONTENT, amountItem.getContent());
-        values.put(COL_DATE, amountItem.getDate());
-        values.put(COL_CARD, amountItem.getCard());
-        values.put(COL_BANK, amountItem.getBank());
-        values.put(COL_CATEGORY, amountItem.getCategory());
-        values.put(COL_AMOUNT, amountItem.getAmount());
+    public long addItem(AmountItem amountItem) {
+        SQLiteDatabase db = null;
+        long uid = -1;
+        try {
+            db = this.getWritableDatabase();
+            ContentValues values = new ContentValues();
+            values.put(COL_CONTENT, amountItem.getContent());
+            values.put(COL_DATE, amountItem.getDate());
+            values.put(COL_CARD, amountItem.getCard());
+            values.put(COL_BANK, amountItem.getBank());
+            values.put(COL_CATEGORY, amountItem.getCategory());
+            values.put(COL_AMOUNT, amountItem.getAmount());
 
-        db.insert(item_TABLE_NAME, null, values);
-        db.close();
-        Toast.makeText(this.context, "아이템 등록", Toast.LENGTH_SHORT).show();
+            // 삽입 후 생성된 UID 값 반환
+            uid = db.insert(item_TABLE_NAME, null, values);
+
+            // UID 값은 삽입된 항목의 고유 ID
+            amountItem.setUid((int) uid);
+            Toast.makeText(this.context, "아이템 등록", Toast.LENGTH_SHORT).show();
+        }catch (Exception e) {
+            Toast.makeText(this.context, "아이템 등록 실패", Toast.LENGTH_SHORT).show();
+        } finally {
+            if (db!= null) {
+                db.close();
+            }
+        }
+        return uid;
     }
 
     // item 정보 업데이트
@@ -127,13 +175,15 @@ public class DBHelper extends SQLiteOpenHelper {
         values.put(COL_CATEGORY, amountItem.getCategory());
         values.put(COL_AMOUNT, amountItem.getAmount());
 
-        return db.update(item_TABLE_NAME, values, COL_CONTENT + "=?", new String[]{amountItem.getContent()});
+        String whereClause =  COL_DATE + "=? AND " + COL_CONTENT + "=? AND " + UID + "=?";
+        String[] whereArgs = {amountItem.getDate(), amountItem.getContent(), String.valueOf(amountItem.getUid())};
+        return db.update(item_TABLE_NAME, values, whereClause, whereArgs);
     }
 
     // item 삭제
     public void deleteItem(AmountItem amountItem) {
         SQLiteDatabase db = this.getWritableDatabase();
-        db.delete(item_TABLE_NAME, COL_CONTENT + "=?" , new String[]{amountItem.getContent()});
+        db.delete(item_TABLE_NAME, COL_CONTENT + "=? AND " + UID + "=?" , new String[]{amountItem.getContent(), String.valueOf(amountItem.getUid())});
         db.close();
     }
 }
